fix: token expireSeconds int -> long

+++ yeah-common/yeah-auth/src/main/java/com/yeshimin/yeahboot/auth/service/JwtService.java
@@ -172,7 +172,7 @@ public class JwtService {
         return payloadVo;
     }
 
-    public Integer getExpireSeconds(String subject) {
+    public Long getExpireSeconds(String subject) {
         return jwtConfigs.get(subject).getExpireSeconds();
     }
 }
